requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
flatbuffers==23.5.26
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
idna==3.6
imageio==2.33.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
onnxruntime==1.16.3
openai==1.6.1
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
pydantic==2.5.3
pydantic_core==2.14.6
PyMatting==1.1.12
pyreadline3==3.4.1
python-dotenv==1.0.0
referencing==0.32.0
rembg==2.0.53
requests==2.31.0
rpds-py==0.16.2
scikit-image==0.22.0
scipy==1.11.4
sniffio==1.3.0
sympy==1.12
tifffile==2023.12.9
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
